Create queues for accounts
This allows you to specify more accounts than you have search workers.

When a search worker fails and reinitializes it's API, it gets a new account.
This is especially useful for upcoming queueing methods where a single worker is responsible for an area.  With this method, if that worker is banned it can switch accounts instead of failing to scan it's assigned area.